Move space-partition exclusion to planner.
Previously the space-partition-exclusion was in the parser.This commit
moves it into the planner and works by pruning the append relation
directly instead of by just adding a qual for the constraint exclusion.
This has several advantages:

1) We can now handle more than one hypertable in a query and in general
make a lot less assumptions.
2) This can now handle optimizations within a view of a hypertable.
3) This does not need to add a qual to the query. This is a performance
improvement because the additional qual does not need to be checked on
each row.

One possible downside is that this method makes more timescale catalog
lookups to do the exclusion. These are not cached and it may increase
planning latency. In small tests this does not seem to be a problem
but we should keep an eye on this and possibly optimize in the future.

Fixes #466.